pyproject.toml
[tool.poetry]
name = "test-api-ecomerce"
version = "0.1.0"
description = ""
authors = ["vanchez98 <iom.boghean@gmail.com>"]
readme = "README.md"
packages = [{include = "test_api_ecomerce"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.6"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
django-filter = "^24.3"
psycopg2 = "^2.9.10"
environ = "^1.0"
django-environ = "^0.12.0"
drf-yasg = "^1.21.8"
drf-friendly-errors-mod = "^0.13.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
